requirements.txt
beautifulsoup4
requests
selenium
google-genai
python-dotenv
google-generativeai
grpcio 
grpc-stubs
grpcio-tools
psycopg2-binary
grpcio-reflection 
grpcio-health-checking 
protobuf 
mypy-protobuf 
types-protobuf 
googleapis-common-protos
pytz
fasttext-wheel
nltk 
sumy